fix: cambio de tipo de columna para evitar error al crear relaciones

+++ database/migrations/2021_03_12_012752_create_propierties_table.php
@@ -15,14 +15,14 @@ class CreateHousesTable extends Migration
     {
         Schema::create('propierties', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->unsignedInteger('user_id');
+            $table->unsignedBigInteger('user_id');
             $table->string('title');
             $table->string('information');
             $table->float('price');
             $table->float('dimension');
-            $table->unsignedInteger('status_id');
-            $table->unsignedInteger('categorie_id');
-            $table->unsignedInteger('city');
+            $table->unsignedBigInteger('status_id');
+            $table->unsignedBigInteger('categorie_id');
+            $table->unsignedBigInteger('city');
             $table->timestamps();
         });
     }
